requirements.txt
numpy>=1.24
pyyaml>=6.0
scipy>=1.10
websockets>=12.0
